+++ src/JsonBuilder.cpp
@@ -4,22 +4,22 @@
 
 #include "JsonBuilder.h"
 
-StringBuffer JsonBuilder::cpuPart(CpuParams cpuParams) {
-    StringBuffer cpuPart;
-    Writer<StringBuffer> writer(cpuPart);
+// StringBuffer JsonBuilder::cpuPart(CpuParams cpuParams) {
+//     StringBuffer cpuPart;
+//     Writer<StringBuffer> writer(cpuPart);
 
-    writer.StartObject();
-    writer.Key("User");
-    writer.Double(cpuParams.getUser());
-    writer.Key("System");
-    writer.Double(cpuParams.getSystem());
-    writer.Key("Wait");
-    writer.Double(cpuParams.getWait());
-    writer.EndObject();
-    cout << cpuPart.GetString() << endl;
+//     writer.StartObject();
+//     writer.Key("User");
+//     writer.Double(cpuParams.getUser());
+//     writer.Key("System");
+//     writer.Double(cpuParams.getSystem());
+//     writer.Key("Wait");
+//     writer.Double(cpuParams.getWait());
+//     writer.EndObject();
+//     cout << cpuPart.GetString() << endl;
 
-    return cpuPart;
-}
+//     return cpuPart;
+// }
 
 StringBuffer JsonBuilder::memoryPart(MemoryParams memoryParams) {
     StringBuffer memory;
@@ -59,7 +59,20 @@ StringBuffer JsonBuilder::allStatus(StatsParam statsParam) {
     rapidjson::Document::AllocatorType& allocator = document.GetAllocator();
 
 // create a rapidjson object type
-    // rapidjson::Value cpu(rapidjson::kObjectType);
+    
+    Value cpuNum(kArrayType);
+    {
+            for(auto d : statsParam.cpuParams){
+            Value cpus(kObjectType);
+            Value frq,model;
+            frq.SetString(StringRef(d.getFrequency().c_str()));
+            model.SetString(StringRef(d.getmodel().c_str()));
+            cpus.AddMember("cores", d.getNumber(), alloc);
+            cpus.AddMember("frequency", frq, alloc);
+            cpus.AddMember("model", model, alloc);
+            cpuNum.PushBack(cpus,alloc);
+            }  
+    }
     // cpu.AddMember("user", statsParam.cpuParams.getUser(), allocator);
     // cpu.AddMember("system",statsParam.cpuParams.getSystem() , allocator);
     // cpu.AddMember("wait", statsParam.cpuParams.getWait(), allocator);
@@ -70,24 +83,32 @@ StringBuffer JsonBuilder::allStatus(StatsParam statsParam) {
             cpu.PushBack(statsParam.generalParams.getcpuUsed(),allocator);                     
         }
 
+    Value hdd(kArrayType);
+        {      
+            hdd.PushBack(statsParam.generalParams.getHddUsage(),allocator);            
+            hdd.PushBack(statsParam.generalParams.getHddUsed(),allocator);                     
+        }
+
     rapidjson::Value loadAvrage(rapidjson::kObjectType);
     loadAvrage.AddMember("user", statsParam.loadAverage.getUser(), allocator);
     loadAvrage.AddMember("system",statsParam.loadAverage.getSystem() , allocator);
     loadAvrage.AddMember("wait", statsParam.loadAverage.getWait(), allocator);
 
-    rapidjson::Value memory(rapidjson::kObjectType);
-    memory.AddMember("used", statsParam.memoryParams.getUsed(), allocator);
-    memory.AddMember("usage",statsParam.memoryParams.getUsage() , allocator);
+    //apidjson::Value memory(rapidjson::kObjectType);
+    Value memory(kArrayType);
+    {
+        memory.PushBack(statsParam.memoryParams.getUsage(), allocator);
+        memory.PushBack(statsParam.memoryParams.getUsed() , allocator);
 
-    rapidjson::Value swap(rapidjson::kObjectType);
-    swap.AddMember("used", statsParam.swapParams.getUsed(), allocator);
-    swap.AddMember("usage",statsParam.swapParams.getUsage() , allocator);
+    }    
+
+    //rapidjson::Value swap(rapidjson::kObjectType);
+    Value swap(kArrayType);
+    swap.PushBack(statsParam.swapParams.getUsage(), allocator);
+    swap.PushBack(statsParam.swapParams.getUsed() , allocator);
 
     rapidjson::Value system(rapidjson::kObjectType);
     Value osVersion;
-    // string versionCommand = "guix --version |grep guix | cut -f 4 -d \" \"";
-    // string version =UTILS::COMMAND::Execute(versionCommand.c_str());
-    // version.erase(std::remove(version.begin(), version.end(), '\n'), version.end());
     osVersion.SetString(statsParam.generalParams.getVersion().c_str(),alloc);
     Value sys;    
     sys.SetString(StringRef(statsParam.generalParams.getSystem().c_str()));
@@ -97,6 +118,7 @@ StringBuffer JsonBuilder::allStatus(StatsParam statsParam) {
     system.AddMember("cpu", cpu, allocator);
     system.AddMember("memory",memory , allocator);
     system.AddMember("swap",swap , allocator);
+    system.AddMember("hdd",hdd , allocator);
     rapidjson::Value Hardware(rapidjson::kObjectType);
         {
         Value disks(kArrayType);
@@ -111,6 +133,8 @@ StringBuffer JsonBuilder::allStatus(StatsParam statsParam) {
             disk.AddMember("usage", d.getUsage(), alloc);
             disks.PushBack(disk, alloc);
             }
+        Hardware.AddMember("memory", statsParam.memoryParams.getTotal(), alloc);
+        Hardware.AddMember("cpu", cpuNum, alloc);
         Hardware.AddMember("disks", disks, alloc);
     }